+++ examples/dialogue_bot/src/transitions.rs
@@ -2,19 +2,23 @@ use crate::states::{
     Dialogue, Receive10x5AnswerState, ReceiveDaysOfWeekState,
     ReceiveGandalfAlternativeNameState, StartState,
 };
+
 use teloxide::prelude::*;
+use teloxide_macros::teloxide;
 
 pub type Out = TransitionOut<Dialogue>;
 
-pub async fn start(cx: TransitionIn, state: StartState) -> Out {
+#[teloxide(transition)]
+async fn start(state: StartState, cx: TransitionIn) -> Out {
     cx.answer_str("Let's start our test! How many days per week are there?")
         .await?;
     next(state.up())
 }
 
-pub async fn receive_days_of_week(
-    cx: TransitionIn,
+#[teloxide(transition)]
+async fn receive_days_of_week(
     state: ReceiveDaysOfWeekState,
+    cx: TransitionIn,
 ) -> Out {
     match cx.update.text().map(str::parse) {
         Some(Ok(ans)) if ans == 7 => {
@@ -28,9 +32,10 @@ pub async fn receive_days_of_week(
     }
 }
 
-pub async fn receive_10x5_answer(
-    cx: TransitionIn,
+#[teloxide(transition)]
+async fn receive_10x5_answer(
     state: Receive10x5AnswerState,
+    cx: TransitionIn,
 ) -> Out {
     match cx.update.text().map(str::parse) {
         Some(Ok(ans)) if ans == 50 => {
@@ -44,9 +49,10 @@ pub async fn receive_10x5_answer(
     }
 }
 
-pub async fn receive_gandalf_alternative_name(
-    cx: TransitionIn,
+#[teloxide(transition)]
+async fn receive_gandalf_alternative_name(
     state: ReceiveGandalfAlternativeNameState,
+    cx: TransitionIn,
 ) -> Out {
     match cx.update.text() {
         Some(ans) if ans == "Mithrandir" => {